+++ crates/core/moon/src/lib.rs
@@ -83,7 +83,6 @@ pub async fn load_workspace_from(path: &Path) -> miette::Result<Workspace> {
                 &workspace.toolchain_config.typescript,
                 &workspace.root,
                 Arc::clone(&workspace.proto_env),
-                workspace.toolchain_config.node.is_some(),
             )),
         );
     }
@@ -108,7 +107,6 @@ pub async fn load_workspace_from(path: &Path) -> miette::Result<Workspace> {
                 &workspace.toolchain_config.typescript,
                 &workspace.root,
                 Arc::clone(&workspace.proto_env),
-                workspace.toolchain_config.bun.is_some(),
             )),
         );
     }