+++ crates/core/platform/src/platform.rs
@@ -2,8 +2,8 @@ use async_trait::async_trait;
 use moon_action_context::ActionContext;
 use moon_common::Id;
 use moon_config::{
-    DependencyConfig, HasherConfig, PlatformType, ProjectConfig, ProjectsAliasesMap,
-    ProjectsSourcesMap, TasksConfigsMap, UnresolvedVersionSpec,
+    DependencyConfig, HasherConfig, PlatformType, ProjectConfig, ProjectsAliasesList,
+    ProjectsSourcesList, TasksConfigsMap, UnresolvedVersionSpec,
 };
 use moon_hash::ContentHasher;
 use moon_platform_runtime::{Runtime, RuntimeReq};
@@ -39,8 +39,8 @@ pub trait Platform: Send + Sync {
     /// map of projects that are unique to the platform's ecosystem.
     fn load_project_graph_aliases(
         &mut self,
-        projects_map: &ProjectsSourcesMap,
-        aliases_map: &mut ProjectsAliasesMap,
+        projects_list: &ProjectsSourcesList,
+        aliases_list: &mut ProjectsAliasesList,
     ) -> miette::Result<()> {
         Ok(())
     }