+++ nextgen/project-graph/src/project_events.rs
@@ -1,6 +1,6 @@
 use moon_common::path::WorkspaceRelativePathBuf;
 use moon_common::Id;
-use moon_config::{DependencyConfig, TaskConfig};
+use moon_config::{DependencyConfig, ProjectsAliasesList, ProjectsSourcesList, TaskConfig};
 use rustc_hash::FxHashMap;
 use starbase_events::Event;
 use std::path::PathBuf;
@@ -9,13 +9,13 @@ use std::path::PathBuf;
 
 #[derive(Debug)]
 pub struct ExtendProjectGraphEvent {
-    pub sources: FxHashMap<Id, WorkspaceRelativePathBuf>,
+    pub sources: ProjectsSourcesList,
     pub workspace_root: PathBuf,
 }
 
 #[derive(Debug, Default)]
 pub struct ExtendProjectGraphData {
-    pub aliases: FxHashMap<String, Id>,
+    pub aliases: ProjectsAliasesList,
 }
 
 impl Event for ExtendProjectGraphEvent {