+++ packages/cli/CHANGELOG.md
@@ -12,10 +12,20 @@
 
 ## Unreleased
 
+#### 🚀 Updates
+
+- Reworked duplicate project ID/alias detection to be more accurate. Will also now error instead of
+  warn when a duplicate is detected.
+  - For aliases, the error can be disabled with the new `experiments.strictProjectAliases` setting
+    in `.moon/workspace.yml`.
+  - For project IDs, the error can not be disabled, as conflicting IDs will cause issues with the
+    project graph.
+
 #### 🐞 Fixes
 
 - Fixed glob based project locating to not log warnings when a file is found and it starts with `.`
   (ignore dotfiles).
+- Fixed project aliases potentially overwriting a project with the same name.
 
 ## 1.18.4